Logger improvements
* Ensure logger can log exception objects
* Remove logging output in tests (for now)
